Fix models

+++ Korenko HM 24/Korenko HM 24/Models/CurrentWeatherData.swift
@@ -50,7 +50,7 @@ struct CurrentWeaterData: Codable {
     let windDeg: Int
     let windGust: Double
     let weather: [WeatherElement]
-    let pop: Int?
+    let pop: Double?
 
     enum CodingKeys: String, CodingKey {
         case dt, sunrise, sunset, temp
